+++ src/main/java/model/piece/Cannon.java
@@ -1,24 +1,24 @@
-package model;
+package model.piece;
 
-import static model.Movement.DOWN;
-import static model.Movement.LEFT;
-import static model.Movement.RIGHT;
+import static model.Movement.*;
 
 import java.util.ArrayList;
 import java.util.List;
+import model.Movement;
+import model.Team;
+import model.position.Position;
 
+public class Cannon extends Piece {
 
-public class Chariot extends Piece {
+    private final List<Movement> movements = List.of(UP, DOWN, LEFT, RIGHT);
 
-    private final List<Movement> movements = List.of(Movement.UP, DOWN, LEFT, RIGHT);
-
-    public Chariot(Team team) {
+    public Cannon(Team team) {
         super(team);
     }
 
     @Override
     public boolean isCannon() {
-        return false;
+        return true;
     }
 
     @Override
@@ -26,7 +26,7 @@ public class Chariot extends Piece {
         return calculatePositionOfMovement(departure, arrival);
     }
 
-    // TODO: Depth 분리
+    //TODO : Depth 분리
     private List<Position> calculatePositionOfMovement(Position departure, Position arrival) {
         for (Movement movement : movements) {
             List<Position> temporaryPosition = new ArrayList<>();
@@ -45,8 +45,8 @@ public class Chariot extends Piece {
     @Override
     public String toString() {
         if (getTeam() == Team.RED) {
-            return "車";
+            return "包";
         }
-        return "차";
+        return "포";
     }
 }
