+++ src/main/java/model/piece/Chariot.java
@@ -1,21 +1,27 @@
-package model;
+package model.piece;
 
-import static model.Movement.*;
+import static model.Movement.DOWN;
+import static model.Movement.LEFT;
+import static model.Movement.RIGHT;
 
 import java.util.ArrayList;
 import java.util.List;
+import model.Movement;
+import model.Team;
+import model.position.Position;
 
-public class Cannon extends Piece {
 
-    private final List<Movement> movements = List.of(UP, DOWN, LEFT, RIGHT);
+public class Chariot extends Piece {
 
-    public Cannon(Team team) {
+    private final List<Movement> movements = List.of(Movement.UP, DOWN, LEFT, RIGHT);
+
+    public Chariot(Team team) {
         super(team);
     }
 
     @Override
     public boolean isCannon() {
-        return true;
+        return false;
     }
 
     @Override
@@ -23,7 +29,7 @@ public class Cannon extends Piece {
         return calculatePositionOfMovement(departure, arrival);
     }
 
-    //TODO : Depth 분리
+    // TODO: Depth 분리
     private List<Position> calculatePositionOfMovement(Position departure, Position arrival) {
         for (Movement movement : movements) {
             List<Position> temporaryPosition = new ArrayList<>();
@@ -42,8 +48,8 @@ public class Cannon extends Piece {
     @Override
     public String toString() {
         if (getTeam() == Team.RED) {
-            return "包";
+            return "車";
         }
-        return "포";
+        return "차";
     }
 }
