+++ src/main/java/model/piece/Elephant.java
@@ -1,19 +1,26 @@
-package model;
+package model.piece;
 
 import static model.Movement.*;
 
 import java.util.ArrayList;
 import java.util.List;
+import model.Movement;
+import model.Team;
+import model.position.Position;
 
-public class Horse extends Piece {
+public class Elephant extends Piece {
 
     private final List<List<Movement>> movements = List.of(
-        List.of(UP, UP_UP_LEFT), List.of(UP, UP_UP_RIGHT),
-        List.of(DOWN, DOWN_DOWN_LEFT), List.of(DOWN, DOWN_DOWN_RIGHT),
-        List.of(LEFT, LEFT_DOWN_LEFT), List.of(LEFT, LEFT_UP_LEFT),
-        List.of(RIGHT, RIGHT_DOWN_RIGHT), List.of(RIGHT, RIGHT_UP_RIGHT));
+        List.of(UP, UP_UP_LEFT, UP_UP_LEFT_UP_LEFT),
+        List.of(UP, UP_UP_RIGHT, UP_UP_RIGHT_UP_RIGHT),
+        List.of(DOWN, DOWN_DOWN_LEFT, DOWN_DOWN_LEFT_DOWN_LEFT),
+        List.of(DOWN, DOWN_DOWN_RIGHT, DOWN_DOWN_RIGHT_DOWN_RIGHT),
+        List.of(LEFT, LEFT_DOWN_LEFT, LEFT_DOWN_LEFT_DOWN_LEFT),
+        List.of(LEFT, LEFT_UP_LEFT, LEFT_UP_LEFT_UP_LEFT),
+        List.of(RIGHT, RIGHT_DOWN_RIGHT, RIGHT_DOWN_RIGHT_DOWN_RIGHT),
+        List.of(RIGHT, RIGHT_UP_RIGHT, RIGHT_UP_RIGHT_UP_RIGHT));
 
-    public Horse(Team team) {
+    public Elephant(Team team) {
         super(team);
     }
 
@@ -43,7 +50,7 @@ public class Horse extends Piece {
         temporaryPosition.add(temporaryMoves);
     }
 
-    private void addMoveByDeparture(Position departure, Movement movement, List<Position> temporaryMoves) {
+    private static void addMoveByDeparture(Position departure, Movement movement, List<Position> temporaryMoves) {
         if (!departure.canMove(movement)) {
             return;
         }
@@ -52,7 +59,6 @@ public class Horse extends Piece {
 
     private List<Position> findArrivalDirection(Position arrival, List<List<Position>> temporaryPosition) {
         return temporaryPosition.stream()
-            .filter(positions -> !positions.isEmpty())
             .filter(positions -> positions.getLast().equals(arrival))
             .findFirst()
             .orElseThrow(() -> new IllegalArgumentException("해당 위치로 이동할 수 없습니다."));
@@ -61,8 +67,8 @@ public class Horse extends Piece {
     @Override
     public String toString() {
         if (getTeam() == Team.RED) {
-            return "馬";
+            return "象";
         }
-        return "마";
+        return "상";
     }
 }
