+++ src/main/java/model/piece/Horse.java
@@ -1,23 +1,22 @@
-package model;
+package model.piece;
 
 import static model.Movement.*;
 
 import java.util.ArrayList;
 import java.util.List;
+import model.Movement;
+import model.Team;
+import model.position.Position;
 
-public class Elephant extends Piece {
+public class Horse extends Piece {
 
     private final List<List<Movement>> movements = List.of(
-        List.of(UP, UP_UP_LEFT, UP_UP_LEFT_UP_LEFT),
-        List.of(UP, UP_UP_RIGHT, UP_UP_RIGHT_UP_RIGHT),
-        List.of(DOWN, DOWN_DOWN_LEFT, DOWN_DOWN_LEFT_DOWN_LEFT),
-        List.of(DOWN, DOWN_DOWN_RIGHT, DOWN_DOWN_RIGHT_DOWN_RIGHT),
-        List.of(LEFT, LEFT_DOWN_LEFT, LEFT_DOWN_LEFT_DOWN_LEFT),
-        List.of(LEFT, LEFT_UP_LEFT, LEFT_UP_LEFT_UP_LEFT),
-        List.of(RIGHT, RIGHT_DOWN_RIGHT, RIGHT_DOWN_RIGHT_DOWN_RIGHT),
-        List.of(RIGHT, RIGHT_UP_RIGHT, RIGHT_UP_RIGHT_UP_RIGHT));
+        List.of(UP, UP_UP_LEFT), List.of(UP, UP_UP_RIGHT),
+        List.of(DOWN, DOWN_DOWN_LEFT), List.of(DOWN, DOWN_DOWN_RIGHT),
+        List.of(LEFT, LEFT_DOWN_LEFT), List.of(LEFT, LEFT_UP_LEFT),
+        List.of(RIGHT, RIGHT_DOWN_RIGHT), List.of(RIGHT, RIGHT_UP_RIGHT));
 
-    public Elephant(Team team) {
+    public Horse(Team team) {
         super(team);
     }
 
@@ -47,7 +46,7 @@ public class Elephant extends Piece {
         temporaryPosition.add(temporaryMoves);
     }
 
-    private static void addMoveByDeparture(Position departure, Movement movement, List<Position> temporaryMoves) {
+    private void addMoveByDeparture(Position departure, Movement movement, List<Position> temporaryMoves) {
         if (!departure.canMove(movement)) {
             return;
         }
@@ -56,6 +55,7 @@ public class Elephant extends Piece {
 
     private List<Position> findArrivalDirection(Position arrival, List<List<Position>> temporaryPosition) {
         return temporaryPosition.stream()
+            .filter(positions -> !positions.isEmpty())
             .filter(positions -> positions.getLast().equals(arrival))
             .findFirst()
             .orElseThrow(() -> new IllegalArgumentException("해당 위치로 이동할 수 없습니다."));
@@ -64,8 +64,8 @@ public class Elephant extends Piece {
     @Override
     public String toString() {
         if (getTeam() == Team.RED) {
-            return "象";
+            return "馬";
         }
-        return "상";
+        return "마";
     }
 }
